requirements.txt
# Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-CORS==4.0.0
Flask-Injector==0.15.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API Documentation
flask-smorest==0.42.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
